Wordpress API connection template
This is a wordpress template to connect to the API and display results for SightMap.  To activate, simply place this in your wordpress theme folder, and on a page you would like to show these results, select the page template as "Numbers Page Sightmap".

If you have any questions please let me know.

Thanks
Mike